agregar funcionalidad para que el forumulario de registro se deslize
