+++ client/src/App.js
@@ -17,7 +17,6 @@ import Registrarse from './features/autenticacion/Registrarse';
 import useIsAuthenticated from './hooks/isAuthenticated';
 import ResetPassword from './features/autenticacion/resetPassword';
 import RecuperarCuenta from './features/autenticacion/RecuperarCuenta';
-import Registrarse_2 from './features/autenticacion/Registrarse_2';
 
 
 function App() {
@@ -54,7 +53,6 @@ function App() {
           <Route path="/registrar" element={<Registrarse/>}/>
           <Route path="/recuperar" element={<RecuperarCuenta/>}/>
           <Route path='/*' element={<Default/>}/>
-          <Route path='registrarse' element={<Registrarse_2/>}/>
         </Routes>
         <Alert/>
       </Router>
