+++ client/src/features/autenticacion/RecuperarCuenta.css
@@ -35,45 +35,6 @@
     align-items: center;
 }
 
-.form-indicators{
-    position: absolute;
-    width: 5rem;
-    height: 1rem;
-    display: flex;
-    justify-content: center;
-    gap: 1rem;
-    top: 0;
-}
-
-.form-indicators .indicator{
-    width: 1.2em;
-    height: 1.2rem;
-    border-radius: 50%;
-    border: 2px solid var(--gris);
-}
-
-.form-indicators .indicator.active{
-    background-color: black;
-}
-
-.form-registro{
-    position: relative;
-    width: 100%;
-    height: 80%;
-    overflow-x: hidden;
-    /* display: flex;
-    align-items: center;
-    overflow-x: visible; */
-}
-
-.form-registro .form-carousel{
-    width: 200%;
-    height: 100%;
-    display: flex;
-    /* justify-content: space-around; */
-    transition: 1s;
-    transform: translateX(50%);
-}
 
 .form-itema{
     /* position: relative; */
